refactor: renamed lib to internal package
