+++ internal/model.go
@@ -1,4 +1,4 @@
-package lib
+package internal
 
 type NumberResponse struct {
 	Number int64 `json:"number"`
